# is_dask_collection: micro optimization