In #6096 I realized that `DuckArrayModule("dask")` is called a lot in our tests - 145'835 times. Most of those are from `is_dask_collection` (`is_duck_dask_array`)  This change avoids that the instance needs to be built every time.

```python
import xarray as xr

%timeit xr.core.pycompat.DuckArrayModule("dask").available
%timeit xr.core.pycompat.dsk.available
```

```
18.9 µs ± 97.7 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
77.1 ns ± 1.22 ns per loop (mean ± std. dev. of 7 runs, 10000000 loops each)
```

Which leads to an incredible speed up of our tests of about 2.7 s :grin: ((18.9 - 0.0771) * 145835 / 1e6).